Disable GPG key unlocking with keychain
There's some issue with the GPG "pinentry-mode" where it asks for the
password through an ncurses interface rather than through a GTK
interface. The ncurses interface also displays incorrectly in the
terminal when keychain starts it. The result is an annoying disruption
when keychain is initializing rather than a smooth process of entering
all the passphrases in GTK dialogs.