Unify irg and vrg commands into one
The new version of irg allows you to toggle back and forth between
ripgrep mode and fzf mode.  Ideally, toggling to a different mode would
push the query value onto a variable which would then get popped when
toggling back.  This may be possible but it's a feature for another day.