Adjust the order of some mappings in ideavimrc
When I create a quick reference card, I have to move the mnemonic
comment down to the right of the first line in the section.  If that
line is too long, I can't fit the comment.  So this commit moves shorter
mappings to the top of their sections so I'll have room.