Use --locked with cargo install
"The --locked flag can be used to force Cargo to use the packaged
Cargo.lock file if it is available. This may be useful for ensuring
reproducible builds, to use the exact same set of dependencies that were
available when the package was published. It may also be useful if a
newer version of a dependency is published that no longer builds on your
system, or has other problems. The downside to using --locked is that
you will not receive any fixes or updates to any dependency."